+++ src/links.rs
@@ -26,4 +26,7 @@ pub mod users {
     pub const FRIENDSHIP_UPDATE: &'static str = "https://api.twitter.com/1.1/friendships/update.json";
     pub const FRIENDS_NO_RETWEETS: &'static str = "https://api.twitter.com/1.1/friendships/no_retweets/ids.json";
     pub const FRIENDSHIP_LOOKUP: &'static str = "https://api.twitter.com/1.1/friendships/lookup.json";
+    pub const BLOCK: &'static str = "https://api.twitter.com/1.1/blocks/create.json";
+    pub const UNBLOCK: &'static str = "https://api.twitter.com/1.1/blocks/destroy.json";
+    pub const REPORT_SPAM: &'static str = "https://api.twitter.com/1.1/users/report_spam.json";
 }